more shuffeling
