+++ pkg/backend/v1/dqlite/backend.go
@@ -3,13 +3,12 @@ package dqlite
 import (
 	"context"
 
-	"github.com/canonical/k8s-dqlite/pkg/backend/v1/internal/backend"
-	"github.com/canonical/k8s-dqlite/pkg/backend/v1/sqlite"
+	internal "github.com/canonical/k8s-dqlite/pkg/backend/v1/internal/backend"
 	"github.com/canonical/k8s-dqlite/pkg/limited"
 )
 
 type BackendConfig struct {
-	backend.Config
+	limited.Config
 	//DriverConfig is the dqlite driver config
 	DriverConfig *DriverConfig
 }
@@ -20,10 +19,10 @@ func NewBackend(ctx context.Context, config *BackendConfig) (limited.Backend, er
 		return nil, err
 	}
 
-	return sqlite.Backend{
+	return internal.Backend{
 		Config:        config.Config,
 		Driver:        driver,
 		Notify:        make(chan int64, 100),
-		WatcherGroups: make(map[*sqlite.WatcherGroup]*sqlite.WatcherGroup),
+		WatcherGroups: make(map[*internal.WatcherGroup]*internal.WatcherGroup),
 	}, nil
 }
