+++ pkg/backend/v1/internal/backend/utils.go
@@ -1,4 +1,4 @@
-package backend
+package internal
 
 import (
 	"bytes"
