+++ pkg/backend/v2/internal/backend/utils.go
@@ -1,4 +1,4 @@
-package sqlitev2
+package internal
 
 import (
 	"bytes"
