+++ pkg/backend/v2/sqlite/driver.go
@@ -1,4 +1,4 @@
-package sqlitev2
+package sqlite
 
 import (
 	"context"
